datapath: split datapath mode option values into own package
These are used in several places where a full import of pkg/option (and
thus all its dependencies) can be avoided because only the
DatapathModeVeth and DatapathModeIpvlan constants are used.

Together with the following commits, this helps to reduce the binary
size of cilium-cni and cilium-docker.
