+++ pkg/datapath/option/option.go
@@ -1,0 +1,25 @@
+// Copyright 2020 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package option
+
+// Available options for datapath mode.
+const (
+	// DatapathModeVeth specifies veth datapath mode (i.e. containers are
+	// attached to a network via veth pairs).
+	DatapathModeVeth = "veth"
+
+	// DatapathModeIpvlan specifies ipvlan datapath mode.
+	DatapathModeIpvlan = "ipvlan"
+)
